Develop into master

+++ src/main/java/org/elevenfiftyconsulting/repositories/ShoppingListItemRepository.java
@@ -3,6 +3,8 @@ package org.elevenfiftyconsulting.repositories;
 import org.elevenfiftyconsulting.beans.ShoppingListItem;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ShoppingListItemRepository extends CrudRepository<ShoppingListItem, Integer> {
+public interface ShoppingListItemRepository extends CrudRepository<ShoppingListItem, Long> {
 
+	
+	
 }
